drivers: pinctrl: rzv2m: fix the warning of pfc
This commit to modify source code pinctrl for RZV2M
board, and avoid warning while building kernel.

+++ drivers/pinctrl/rzv2m/pfc.h
@@ -79,7 +79,7 @@
 #define RZV2M_PINCTRL(pin_name, port, pf, config)	\
 	{					\
 		.pin = RZV2M_PIN(port, pf),	\
-		.configs = config,		\
+		.configs = (unsigned short)(config),		\
 		.name = pin_name,		\
 	}
